requirements.txt
pandas 
scikit-allel  
tensorboard 
torch 
wandb
optuna
scipy

